# contracts/openapi/openapi_v1.0.6_enveloped.yaml
openapi: 3.0.3
info:
  title: Hourglass Rentals API
  version: 1.0.6
  description: '面向西班牙（Spain）上线的租赁管理平台 API（Admin / Owner / Tenant + Public Website + CRM）。

    本文件是字段级接口“唯一真相源（Source of Truth）”。前后端均以此为准。


    约定：

    - 所有需要翻译展示的文本字段同时保留原文（*_original）并提供翻译缓存（*_i18n）与展示字段（*_display）。

    - 服务端根据用户偏好与请求头 Accept-Language 解析 viewerLanguage，并填充 *_display。

    - 所有时间使用 ISO 8601，货币默认 EUR。



    ---


    **M0 冻结说明（2025-12-24）**

    - 生产统一使用 `/v1`（server url 已带 `/v1`）；原 `/api` 仅作为短期兼容（可由网关 rewrite）。

    - 统一响应 Envelope：成功响应为 `{requestId,timestamp,data}`；错误响应为 `{requestId,timestamp,error}`。



    ### M0-FULL 补丁说明（v1.0.2）

    - 补齐：/health、/invitations、/gdpr、property images upload

    - 安全：补齐通用错误码（401/403/404/429/500）与 Cookie+CSRF 口径

    - 兼容：/api → /v1 下线计划见 `00_RELEASE/API_兼容与下线计划.md`



    ### M0-SUPER 补丁说明（v1.0.3）


    - 新增：`GET /v1/gdpr/jobs/{jobId}`（GDPR job 查询，便于前端轮询）

    - 新增：`POST /v1/auth/refresh`、`POST /v1/auth/logout`（Cookie 会话刷新/退出）

    - 安全：补充 refresh cookie 的可执行表达（`refreshCookie` security scheme）



    v1.0.4（M0-Ultimate）变更摘要：

    - GDPR：新增历史任务列表/筛选接口 `GET /gdpr/jobs`（支持 type/status/target 过滤 + cursor 分页）。

    - 并发：引入 If-Match / ETag 乐观锁约定（rowVersion 全链路），写接口可选强制门禁（412/428）。


    补充：ErrorResponse.error.code 已固定为 ErrorCode 枚举（前后端统一）。


    【v1.0.6 重要变更（M0修正）】

    - 金额字段：对外统一使用字符串（MoneyString），避免 JSON number 精度问题；落库使用 NUMERIC(18,2)。

    - Document 新增 scanStatus 字段（UPLOADED/SCANNING/READY/BLOCKED/FAILED）。

...

    '
  contact:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '建议使用 Google Identity Platform / Firebase Auth / Auth0 等签发 JWT；API 侧只做校验与 RBAC。


        ⚠️ M0-FULL：浏览器端推荐使用 HttpOnly Cookie 会话（见 sessionCookie + CSRF），Bearer 仅建议用于服务到服务或本地调试。'
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly Secure Session Cookie（浏览器端默认认证方式）。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF 防护 token（对所有写接口 POST/PUT/PATCH/DELETE 必须）。
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh
      description: HttpOnly Secure Refresh Cookie（用于会话刷新/轮换）。
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        example: zh-CN,es;q=0.9,en;q=0.8
      description: 期望返回的展示语言。服务端可结合用户偏好与此头决定 viewerLanguage。
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: 游标分页（不透明字符串，由服务端生成）。
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: 乐观锁：客户端从响应头 ETag 获取并在更新/状态变更/删除时通过 If-Match 回传。推荐格式：W/"{rowVersion}".
  schemas:
    ErrorCode:
      type: string
      description: 统一错误码枚举（用于 ErrorResponse.error.code）。前后端/多端必须保持一致。
      enum:
      - VALIDATION_ERROR
      - BAD_REQUEST
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - CONFLICT
      - PRECONDITION_REQUIRED
      - PRECONDITION_FAILED
      - RATE_LIMITED
      - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required:
      - error
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          description: 请求追踪 ID（用于排障）
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: 服务端生成的响应时间戳
...
    DocumentVisibility:
      type: string
      enum:
      - INTERNAL
      - OWNER
      - TENANT
      - BOTH
    DocumentDocType:
      type: string
      description: 文件分类（可扩展）。
      enum:
      - Other
      - TicketAttachment
      - InvoiceProof
      - LeaseContract
      - LeaseContractSigned
      - ExpenseReceipt
      - DepositProof
      - ApplicantDoc
      - ID_DOC
      - PAYSLIPS
      - BANK_STATEMENTS
      - WORK_CONTRACT
      - PropertyImage
    StorageProvider:
      type: string
      enum:
      - GCS
      - R2
    Document:
      type: object
      required:
      - id
      - name
      - docType
      - visibility
      - mime
      - size
      - createdAt
      - scanStatus
      properties:
        id:
          type: string
        name:
          type: string
        docType:
          $ref: '#/components/schemas/DocumentDocType'
        expiresAt:
          type: string
          format: date
          nullable: true
        visibility:
          $ref: '#/components/schemas/DocumentVisibility'
        ownerId:
          type: string
          nullable: true
        tenantId:
          type: string
          nullable: true
        propertyId:
          type: string
          nullable: true
        leaseId:
          type: string
          nullable: true
        invoiceId:
          type: string
          nullable: true
        ticketId:
          type: string
          nullable: true
        entityType:
          type: string
          nullable: true
          description: 例如 documentRequest / application / inventory ...
        entityId:
          type: string
          nullable: true
        mime:
          type: string
        size:
          type: integer
        sha256:
          type: string
          nullable: true
        uploader:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        storageProvider:
          allOf:
          - $ref: '#/components/schemas/StorageProvider'
          nullable: true
          readOnly: true
        bucket:
          type: string
          nullable: true
          readOnly: true
        objectKey:
          type: string
          nullable: true
          readOnly: true
        downloadUrl:
          type: string
          format: uri
          nullable: true
          readOnly: true
          description: 签名下载链接（短期有效）（仅 scanStatus=READY 时返回/可用）
        previewUrl:
          type: string
          format: uri
          nullable: true
          readOnly: true
        dataUrl:
          type: string
          nullable: true
          description: 仅 Demo/原型使用；生产禁止持久化此字段
        rowVersion:
          type: integer
          format: int64
          description: 用于乐观锁的行版本号（与 ETag/If-Match 对应）。
        scanStatus:
          $ref: '#/components/schemas/DocumentScanStatus'
...
    EnvelopeMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          description: 请求追踪 ID（用于排障）
        timestamp:
          type: string
          format: date-time
          description: 服务端生成的响应时间戳
...
    Envelope_Document:
      allOf:
      - $ref: '#/components/schemas/EnvelopeMeta'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/Document'
    Inline_d5ddc47f0f:
      type: array
      items:
        $ref: '#/components/schemas/Document'
    Envelope_Inline_d5ddc47f0f:
      allOf:
      - $ref: '#/components/schemas/EnvelopeMeta'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/Inline_d5ddc47f0f'
...
    DocumentScanStatus:
      type: string
      description: 文件扫描状态机：UPLOADED → SCANNING → READY / BLOCKED；扫描失败为 FAILED。仅 READY 可生成 downloadUrl/previewUrl。
      enum:
      - UPLOADED
      - SCANNING
      - READY
      - BLOCKED
      - FAILED
...
  responses:
    BadRequest400:
      description: 请求参数错误 / 校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 未登录或会话失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden403:
      description: 无权限（或策略要求返回 404 以防止资源枚举）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound404:
      description: 资源不存在（或越权时按 404 处理）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict409:
      description: 并发冲突 / 乐观锁失败（expected_row_version/If-Match）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests429:
      description: 请求过于频繁（限流触发）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError500:
      description: 服务端异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionFailed412:
      description: Precondition Failed（ETag/If-Match 不匹配，资源已被他人修改）
      content:
        application/json:
          schema: &id001
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionRequired428:
      description: Precondition Required（缺少 If-Match；该接口启用强制乐观锁门禁）
      content:
        application/json:
          schema: *id001
  headers:
    ETag:
      description: 资源版本 ETag（由 rowVersion 派生）。用于 If-Match 乐观锁。
      schema:
        type: string
paths:
...
  /documents:
    get:
      tags:
      - Files
      summary: 查询文档列表（按可见性+权限过滤）
      security:
      - sessionCookie: []
      parameters:
      - $ref: '#/components/parameters/AcceptLanguage'
      - name: ownerId
        in: query
        required: false
        schema:
          type: string
      - name: tenantId
        in: query
        required: false
        schema:
          type: string
      - name: leaseId
        in: query
        required: false
        schema:
          type: string
      - name: invoiceId
        in: query
        required: false
        schema:
          type: string
      - name: ticketId
        in: query
        required: false
        schema:
          type: string
      - name: entityType
        in: query
        required: false
        schema:
          type: string
      - name: entityId
        in: query
        required: false
        schema:
          type: string
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Inline_d5ddc47f0f'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
      x-rate-limit-tier: STANDARD
      x-error-strategy: 403/404
      operationId: getDocuments
      description: 查询文档列表（按可见性+权限过滤）
  /documents/{docId}:
    get:
      tags:
      - Files
      summary: 获取文档元数据（含短期 downloadUrl）
      security:
      - sessionCookie: []
      parameters:
      - name: docId
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Document'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
      x-rate-limit-tier: STANDARD
      x-error-strategy: 403/404
      operationId: getDocumentsByDocId
      description: 获取文档元数据（含短期 downloadUrl）
    patch:
      tags:
      - Files
      summary: 更新文档元数据（可见性/到期日等）
      security:
      - sessionCookie: []
        csrfToken: []
      parameters:
      - name: docId
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/AcceptLanguage'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                docType:
                  $ref: '#/components/schemas/DocumentDocType'
                expiresAt:
                  type: string
                  format: date
                  nullable: true
                visibility:
                  $ref: '#/components/schemas/DocumentVisibility'
              additionalProperties: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Document'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '409':
          $ref: '#/components/responses/Conflict409'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
        '412':
          $ref: '#/components/responses/PreconditionFailed412'
        '428':
          $ref: '#/components/responses/PreconditionRequired428'
      x-rate-limit-tier: HIGH_RISK
      x-audit: true
      x-audit-action: DOCUMENT_UPDATE
      x-error-strategy: 403/404
      description: 并发控制：该接口启用 If-Match 乐观锁（基于 rowVersion/ETag）。
      operationId: patchDocumentsByDocId
  /documents/{docId}/download:
    get:
      tags:
      - Files
      summary: 获取下载链接（或 302 重定向）
      security:
      - sessionCookie: []
      parameters:
      - name: docId
        in: path
        required: true
        schema:
          type: string
      responses:
        '302':
          description: Redirect to signed URL
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Document'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
        '409':
          $ref: '#/components/responses/Conflict409'
      x-rate-limit-tier: HIGH_RISK
      x-audit: true
      x-audit-action: DOCUMENT_DOWNLOAD
      x-error-strategy: 403/404
      description: 仅当 Document.scanStatus=READY 时可下载/重定向；否则返回 409 Conflict。
      operationId: getDocumentsByDocIdDownload
...